Do not play only example first
Prevents a confusing "video cannot be played" toast on the Sign tab when
the first and only example is broken but the sign video works.

Also autoplay means that you often end up in the middle of the example
video when you switch to the example tab and have to anyway restart
playback (or wait for the loop) to see it from the beginning.