Fix synonym relations
The synonym information appears to be incomplete in the source data.
For example, MISSFALL has two signs, 00214 and 05166, but while
the samma-betydelse page for 05166 lists 00214, the corresponding
page for 00214 also only lists 00214.

 https://teckensprakslexikon.su.se/ord/00214/samma-betydelse
 https://teckensprakslexikon.su.se/ord/05166/samma-betydelse

Fix it up.  This would be needed for tail crawls even if there were
no errors in the source data.