Remove requests_cache
We already needed to work around stuff in request_cache that don't work
as we want it.  It also seems to open and close the database on every
request and write to the database even on every cached request making
things very slow.  Just remove it and add a custom cache.